Reimplementation of Brackman in Python (#12)
The majority of the original Javascript Brackman bot has been reimplemented
in Python, because it's much easier for me to understand.  The `f/sort`, `f/set`
and `f/who` commands work, and Brackman now has an interlock so that if two
people trying to sort the same game, all later `f/sort` commands are ignored
for the time it takes to create the voice channels and find and move the players.

Signed-off-by: Paul Wayper <paulway@mabula.net>